Internal #861: Refactor Window Collection
Move the WindowCollection classes to separate files.
