+++ src/function/window/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_library_unity(duckdb_func_window OBJECT window_executor.cpp
-                  window_segment_tree.cpp)
+add_library_unity(duckdb_func_window OBJECT window_collection.cpp
+                  window_executor.cpp window_segment_tree.cpp)
 
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_window>
